CONTRAIL_BUILD_FROM_SOURCE (#63)
* add an extra variable to indicate containers build mode. by default
we want to avoid packaging stage of rpmbuild to speed up the whole
routine.

* google switched repo tool to the 2 version which requires git version
2 and above. it is unavailable from default yum repos. thus the
build process fails with the error of obsolete git.
switch to using of repo-1. this requies changing of the download url
and repo branch.

* now we perform a substring test to match by name a container of
interest. this leads to false positive matches if 2 container
names share the same substring.
let's switch to exact match to workaround.

* update the way we instantiate the template for the build container
to enable bash variable expansions inside the template.
add CONTRAIL_BUILD_FROM_SOURCE environment to the list for the
build container.