try to improve osxfs performance
The disk i/o performance running Docker containers on MacOS are
slow. Several topics related to this issue can be found.

With this change we set 'delegated' to the volumes mounted so the
changes that are happening in the sandbox during executing processes
are done async. This improve the performance of about 10%. This change
is safe since we are not expected the sandbox to update the source
code itself, only the builds it has created

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>