Add support to macosx (#77)
* add experimental support of MacOS

This will install a limitted number of dependencies from brew thatwill
allow to run the sandbox and continue tohack the code locally.

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>

* add support to configure docker for mac osx

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>

* try to improve osxfs performance

The disk i/o performance running Docker containers on MacOS are
slow. Several topics related to this issue can be found.

With this change we set 'delegated' to the volumes mounted so the
changes that are happening in the sandbox during executing processes
are done async. This improve the performance of about 10%. This change
is safe since we are not expected the sandbox to update the source
code itself, only the builds it has created

Signed-off-by: Sahid Orentino Ferdjaoui <sferdjaoui@juniper.net>
Change-Id: I12e580aed1692228daab3c021e7c326a900e614e